refactor: extract and test shouldShowCulturalSurvey
